Support checkpointing and progress reporting.
Upon receiving a SIGINT signal (e.g. from Ctrl+C), the simulation will
stop and save its state. Re-running with the same seed resumes the
simulation from the saved state.

Upon receiving a SIGUSR1 signal in Linux or SIGBREAK (from Ctrl+Break)
in Windows, a short summary of the current status of the simulation is
outputted.